separate types into base and version specific
