+++ powerfactory_tools/versions/pf2022/interface.py
@@ -22,20 +22,20 @@ import loguru
 import pydantic
 
 from powerfactory_tools.base.constants import BaseUnits
+from powerfactory_tools.base.types import CalculationCommand
 from powerfactory_tools.base.types import Currency
 from powerfactory_tools.base.types import FolderType
 from powerfactory_tools.base.types import MetricPrefix
+from powerfactory_tools.base.types import NetworkExtendedCalcType
 from powerfactory_tools.base.types import PFClassId
+from powerfactory_tools.base.types import ResultExportMode
+from powerfactory_tools.base.types import TimeSimulationNetworkCalcType
+from powerfactory_tools.base.types import TimeSimulationType
+from powerfactory_tools.base.types import UnitSystem
+from powerfactory_tools.base.types import ValidPFValue
 from powerfactory_tools.utils.io import CustomEncoder
 from powerfactory_tools.utils.io import FileType
 from powerfactory_tools.versions.pf2022.data import PowerFactoryData
-from powerfactory_tools.versions.pf2022.types import CalculationCommand
-from powerfactory_tools.versions.pf2022.types import NetworkExtendedCalcType
-from powerfactory_tools.versions.pf2022.types import ResultExportMode
-from powerfactory_tools.versions.pf2022.types import TimeSimulationNetworkCalcType
-from powerfactory_tools.versions.pf2022.types import TimeSimulationType
-from powerfactory_tools.versions.pf2022.types import UnitSystem
-from powerfactory_tools.versions.pf2022.types import ValidPFValue
 
 if t.TYPE_CHECKING:
     from collections.abc import Iterable
